pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "manus-mcp"
version = "0.1.0"
description = "An MCP server that can browse the web, perform search queries, and execute code"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.30.0",
    "pydantic>=2.4.2",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "mcp==1.4.1",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 